Update quirkey.py
Moved PgUp/PgDn to the right keys.
Added a backslash key.
Put in a heartbeat LED output for debugging.
Wrapped the main loop in an exception handler, and cause a reset if an exception happens. Without this, the Quirkey would often crash when connected to a hibernating PC because when a key was pressed the USB transmit would fail.
Before starting the main loop process, the code now waits for all keys to be up (displaying a serial port message and rapidly blinking the heartbeat LED). This also gives you a chance to upload new code if things keep crashing in development...